Added more functions for TI-STAR
TI-STAR now has the following command line / macro functions
available:

/DetSys/det/setTISTARFirstLayerOffset <x> <y> <z> <unit>
/DetSys/det/setTISTARFirstLayerPCBUpperX <value> <unit>
/DetSys/det/setTISTARFirstLayerPCBLowerX <value> <unit>

/DetSys/det/setTISTARSecondLayerOffset <x> <y> <z> <unit>
/DetSys/det/setTISTARSecondLayerPCBUpperX <value> <unit>
/DetSys/det/setTISTARSecondLayerPCBLowerX <value> <unit>

/DetSys/det/setTISTARThirdLayerOffset <x> <y> <z> <unit>
/DetSys/det/setTISTARThirdLayerPCBUpperX <value> <unit>
/DetSys/det/setTISTARThirdLayerPCBLowerX <value> <unit>

The offset is further applied to the position calculated via
the y-distance from the beam, the z-gap between the first two
panels, etc. The PCB upper/lower functions can be used to shift
the silicon panels so that the silicon+PCB chip is centered in
about the origin.

The macro used for building TI-STAR with some of these functions
has also been updated.