Re-organizing TI-STAR geometry code
Re-writing the TI-STAR class geometry so that it is much more
general - i.e. we should just have to specify the dimensions
of each layer (Si dims, PCB dims, and location of Si within
the PCB chip), and can then make TI-STAR with an arbitrary
number of layers of arbitrary chip size/shape.