Added more TI-STAR functions
TI-STAR now has functions to build 2-strip and 4-strip layers. The
relevant commands are as follows:

Add a 2-strip layer to TI-STAR (one panel on either side of the beam):
    /DetSys/det/addTISTAR2StripLayer

Add a 4-strip layer to TI-STAR (two panels on either side of the beam):
    /DetSys/det/addTISTAR4StripLayer

Set the detector to beam axis distance:
    /DetSys/det/setTISTARDistFromBeam <value> <unit>

Set the z-gap distance for a 4-strip layer:
    /DetSys/det/setTISTARGapZ <value> <unit>

Set 1 for si-centered layers, 0 for detector-centered layers:
    /DetSys/det/setTISTARSiCentered <boolean_value>

The TI-STAR.mac run macro has been updated to reflect these changes.