Added the rest of the TRex generators
The other primary event generators that were in the TRex code have
now been modified and implemented w/ Griffinv10. These include:
    - TRexTestSource
    - TRexBeamIn
    - TRexAlphaSource
    - TRexRutherford
When running the simulation w/ these generators, the ntuple w/ the
output is included w/ the g4out.root file and the detector ntuple.
The all seem to work fine, except for TRexRutherford, which issues
an error on each event:

-------- EEEE ------- G4Exception-START -------- EEEE -------
*** G4Exception : TRACK001
      issued by : G4VParticleChange::CheckIt
Step length and/or energy deposit was illegal
*** Event Must Be Aborted ***
-------- EEEE -------- G4Exception-END --------- EEEE -------

My guess is this could be to a setting that I have not included or
is not set to a proper value in the run macro that I am using (the
same as in the TI-STAR macro except w/ the Rutherford generator vs
AngularDistribution).