Shifting the inner layers
Working on lining up the inner 4-strip layer of the TI-STAR detector
so that they line up with the outer two 2-strip layers. This is
currently hard-coded just to test that the math/movement vectors
work correctly, which it looks like they do. Will add a function
so that this can be done via the command line soon.