Basic beam spreading added
Still need to have this done via a data file that is read in, but
the basic beam spreading has been added to TRexBeam. This works by
having the x-y position sampled in a circle w/ a radius that
depends on the z-position in the target. The data file will be in
the form of z-position vs. beam radius.