Beam spread file added
A file with data in the form of z-dist in the target vs. the beam
radius can now be used to generate the primary event verticies.
The first line of the data file is used to specify the number of
z-vs-radius points included in the file, same as for the other
data files that are read in. The z-position is still sampled
linearly, which we might want to change at a later point in time.

The data file is set via:
    /DetSys/miniball/SetBeamSpreadFile <file_name>

Should add a switch to turn this on/off, in case we want to use
just the fBeamWidth value instead, although this might need
tweaking as it samples the x-y positions from a square w/ side
lengths of fBeamWidth, not a circle w/ radius of fBeamWidth.