Added ReactionZ prob. distrib. file
Now the reaction Z value can be sampled from a distribution
that is specified via:
    /DetSys/miniball/SetReactionZDistributionFile <filename>
The file must have the number of entries as the first line,
followed by that number of data points in two columns, the first
being the reactionZ value, the second being the probability. The
total probability does not need to be normalized - thi is done
automatically.