Added PhysListEmStandardNR from TestEm7
The example simulation TestEm7 implements the electromagnetic
physics lists just like the Griffinv10 simulation, so it was easy
just to copy over their modified physics list (PhysListEmStandardNR)
and use it with our simulation. This certainly looks to be the
easiest way to implement the G4ScreenedNuclearRecoil class with our
simulation.