Removing modified QGSP_BIC physics list
Removed modified physics list from dhymers, instead for the nuclear
recoils we will use the physics list PhysListEmStandardNR from the
example simulation TestEm7 for use of G4ScreenedNuclearRecoil proc.
Other small changes to the TI-STAR macro.

Changed the RunAction constructior so that it doesn't need a pointer
to the PrimaryGeneratorAction. I think I added this when first
integrating the TRex-based generators, but it doesn't appear to be
neccesary now. This was when trying to fix a seg-fault when running
the code multi-threaded. This error still occurs, and I think might
be to due with the Build() and BuildForMaster() methods there. Will
look into at a later date...