Vacuum chamber exterior
Added the vacuum chamber exterior, and the TI-STAR layers are now
build inside the vacuum chamber. This might mess up the way the
layers are identified - possibly because the DetectorConstruction
Print() method searches through the daugher volumes in the world
volume, but now the TI-STAR layers are daughter volumes to the
vacuum chamber.