New method for hit tracking w/ TI-STAR
A new hit will be created for each unique track in one of the first
two TI-STAR layers (the pixelated ones). This should allow multiple
hits in the same detector layer which can be added together in the
analysis code (NTuple) if they correspond to the same strip/ring
number.